+++ src/core/models/errors.ts
@@ -1,5 +1,5 @@
 export class ErrorBase extends Error {
-    constructor(name: string, message?: string, ) {
+    constructor(name: string, message?: string,) {
         super(message);
         Object.setPrototypeOf(this, new.target.prototype);
         this.name = name;
@@ -9,5 +9,11 @@ export class ErrorBase extends Error {
 export class ItemNotFoundError extends ErrorBase {
     constructor(message?: string) {
         super("ItemNotFoundError", message);
+    }
+}
+
+export class ArgumentError extends ErrorBase {
+    constructor(message?: string) {
+        super("ArgumentError", message);
     }
 }